Dont delete empty blocks, use nop message instead
Deleting a block also deletes the corresponding viewangles. Probably not
a big deal, as we would usually only delete blocks without a time
message, but in general should be cleaner to not delete.

As blocks with no messages do not seem supported (JoeQuake freezes on
playback), we write a nop message in that case.