Fix finding collected item on previous block
Was using hard coded offset of -1 to look for active item that was
collected previous to collection sound event. It has to be the previous
block that has a time message though, not any of the intermediary blocks
that don't have one. Can happen for example with collections that are
very close together, since collections have a block with just some
prints.

So now properly look for the closest previous frame that does have a
time message in it.