More robust handling of collections on 1st frames
Consider pause status for first active block index. But also accept
collections that happened in the original demo, even if they are before
the first active block index, cause that can apparently happen. That way
we are lenient in accepting what happened in the original, but
conservative in what additional collections we look for during the first
few frames.