Keep track of consumed items in original demo
If due to stats fixing we remove the collection of an item, but then
later want to add the collection of exactly that item, we would get a
warning that we expected the original demo to pickup the item but it did
not. That's because we did not keep track of collectables being consumed
in the original demo, explaining why they were not collected later on.
This was even falsely attributed to coop play, which is wrong, exactly
the same situation can happen in singleplayer.

Now just keeping track of consuming in original, and only checking for
whether original also picked it up if not consumed yet.