Handle collectables that turn into something else
Not quite sure how, but sometimes static collectable entities seem to
transform into something else. In that case the stats code should not
consider it to be able to be picked up by a player anymore.

It's possible that this only happens with the item-disappearing-bug when
colliding with a monster, for example the megahealth on e2m5. This fix
for example was trigerred by a nailgun ammo pack on e2m6 Nightmare next
to the LG bug spot not being there in a demo.