Fix multiple protocols per demo
Fix handling of protocols when there are multiple different ones per
demo.

In 871e9f3e, storing the protocol was moved from ServerInfoMessage to
Demo, which simplified implementing the feature of allowing to write out
a demo with a different protocol than what it was recorded with.
Unfortunately that also meant that for demos with multiple
ServerInfoMessage with different protocols in them (e.g. marathon
demos), the protocol changes would not be tracked correctly.

This fixes that by reverting the protocol storage back to
ServerInfoMessage and reimplementing the procotol override feature
during demo writing accordingly.